#24
- `severityFromStatus` перенесена из `http-client/middleware` в `preset/parts/utils`
