+++ src/http-client/middleware/__test__/logging.test.ts
@@ -1,6 +1,6 @@
 import { AxiosRequestConfig, AxiosDefaults, AxiosResponse } from 'axios';
 import { Next } from 'middleware-axios';
-import { loggingMiddleware, LogMiddlewareHandler, severityFromStatus } from '../logging';
+import { loggingMiddleware, LogMiddlewareHandler } from '../logging';
 
 function getFakeHandler(): LogMiddlewareHandler {
   return {
@@ -117,26 +117,5 @@ describe('loggingMiddleware', () => {
     expect(handler.afterResponse).toBeCalledTimes(1);
     expect(handler.afterResponse).toBeCalledWith({ config, defaults, response });
     expect(handler.onCatch).toBeCalledTimes(0);
-  });
-});
-
-describe('severityFromStatus', () => {
-  it('should works', () => {
-    expect(severityFromStatus(200)).toBe('info');
-    expect(severityFromStatus(201)).toBe('info');
-    expect(severityFromStatus(204)).toBe('info');
-
-    expect(severityFromStatus(300)).toBe('warning');
-    expect(severityFromStatus(302)).toBe('warning');
-    expect(severityFromStatus(400)).toBe('warning');
-    expect(severityFromStatus(404)).toBe('warning');
-    expect(severityFromStatus(422)).toBe('warning');
-    expect(severityFromStatus(499)).toBe('warning');
-
-    expect(severityFromStatus(undefined)).toBe('error');
-    expect(severityFromStatus(100)).toBe('error');
-    expect(severityFromStatus(199)).toBe('error');
-    expect(severityFromStatus(500)).toBe('error');
-    expect(severityFromStatus(503)).toBe('error');
   });
 });
