+++ src/preset/parts/__test__/utils.test.ts
@@ -1,9 +1,8 @@
 import { Env } from '@humanwhocodes/env';
 import { AxiosRequestConfig, AxiosDefaults } from 'axios';
 import { SentryBreadcrumb, SentryError } from '../../../error-tracking';
-import { severityFromStatus } from '../../../http-client/middleware/logging';
 import { Logger } from '../../../logger';
-import { HttpApiHostPool, HttpClientLogging, SagaLogging } from '../utils';
+import { HttpApiHostPool, HttpClientLogging, SagaLogging, severityFromStatus } from '../utils';
 
 const logger: Logger = {
   log: jest.fn(),
@@ -39,6 +38,27 @@ describe('HttpApiHostPool', () => {
     expect(() => {
       pool.get('bar' as any);
     }).toThrow(`Known HTTP API not found by key "bar"`);
+  });
+});
+
+describe('severityFromStatus', () => {
+  it('should works', () => {
+    expect(severityFromStatus(200)).toBe('info');
+    expect(severityFromStatus(201)).toBe('info');
+    expect(severityFromStatus(204)).toBe('info');
+
+    expect(severityFromStatus(300)).toBe('warning');
+    expect(severityFromStatus(302)).toBe('warning');
+    expect(severityFromStatus(400)).toBe('warning');
+    expect(severityFromStatus(404)).toBe('warning');
+    expect(severityFromStatus(422)).toBe('warning');
+    expect(severityFromStatus(499)).toBe('warning');
+
+    expect(severityFromStatus(undefined)).toBe('error');
+    expect(severityFromStatus(100)).toBe('error');
+    expect(severityFromStatus(199)).toBe('error');
+    expect(severityFromStatus(500)).toBe('error');
+    expect(severityFromStatus(503)).toBe('error');
   });
 });
